# services/main/docs/swagger.yaml
definitions:
...
  handlers.CreatePropertyBlockRequest:
    properties:
      description:
        example: Spacious apartment with sea view.
        type: string
      images:
        example:
        - https://example.com/image1.jpg
        - https://example.com/image2.jpg
        items:
          type: string
        type: array
      name:
        example: Luxury Apartment
        maxLength: 100
        minLength: 2
        type: string
      status:
        enum:
        - PropertyBlock.Status.Active
        - PropertyBlock.Status.Maintenance
        - PropertyBlock.Status.Inactive
        example: PropertyBlock.Status.Active
        type: string
    required:
    - name
    - status
    type: object
...
  handlers.UpdatePropertyBlockRequest:
    properties:
      description:
        example: Spacious apartment with sea view.
        type: string
      images:
        example:
        - https://example.com/image1.jpg
        - https://example.com/image2.jpg
        items:
          type: string
        type: array
      name:
        example: Luxury Apartment
        maxLength: 100
        minLength: 2
        type: string
      status:
        enum:
        - PropertyBlock.Status.Active
        - PropertyBlock.Status.Maintenance
        - PropertyBlock.Status.Inactive
        example: PropertyBlock.Status.Active
        type: string
    type: object
...
  lib.HTTPError:
    properties:
      errors:
        properties:
          message:
            type: string
        type: object
    type: object
  lib.HTTPReturnPaginatedMetaResponse:
    properties:
      has_next_page:
        example: true
        type: boolean
      has_previous_page:
        example: false
        type: boolean
      order:
        example: desc
        type: string
      order_by:
        example: created_at
        type: string
      page:
        example: 1
        type: integer
      page_size:
        example: 20
        type: integer
      total:
        example: 100
        type: integer
    type: object
...
  transformations.OutputAdmin:
    properties:
      created_at:
        example: "2023-01-01T00:00:00Z"
        type: string
      email:
        example: admin@example.com
        type: string
      id:
        example: 4fce5dc8-8114-4ab2-a94b-b4536c27f43b
        type: string
      name:
        example: Admin Name
        type: string
      updated_at:
        example: "2023-01-01T00:00:00Z"
        type: string
    type: object
...
  transformations.OutputClient:
    properties:
      address:
        example: 123 Main St, Suite 100
        type: string
      city:
        example: San Francisco
        type: string
      client_application:
        $ref: '#/definitions/transformations.OutputClientApplication'
      client_application_id:
        example: app-1234
        type: string
      country:
        example: US
        type: string
      created_at:
        example: "2023-01-01T00:00:00Z"
        type: string
      id:
        example: 4fce5dc8-8114-4ab2-a94b-b4536c27f43b
        type: string
      latitude:
        example: 37.7749
        type: number
      longitude:
        example: -122.4194
        type: number
      name:
        example: Acme Corp
        type: string
      region:
        example: California
        type: string
      sub_type:
        example: LANDLORD
        type: string
      support_email:
        example: support@somewebiste.com
        type: string
      support_phone:
        example: "+233551235555"
        type: string
      type:
        example: INDIVIDUAL
        type: string
      updated_at:
        example: "2023-01-01T00:00:00Z"
        type: string
      website_url:
        example: https://www.somewebiste.com
        type: string
    type: object
  transformations.OutputClientApplication:
    properties:
      address:
        description: company address or individual home address
        example: 21st Neon Street
        type: string
      approved_by_id:
        example: S90092
        type: string
      city:
        example: Accra
        type: string
      contact_email:
        example: contact@email.com
        type: string
      contact_name:
        example: John Doe
        type: string
      contact_phone_number:
        example: "01234567890"
        type: string
      country:
        example: Ghana
        type: string
      created_at:
        example: "2023-01-01T00:00:00Z"
        type: string
      date_of_birth:
        description: individual specific fields
        example: "2025-01-31"
        type: string
      description:
        example: Taking you to the next level!
        type: string
      id:
        example: 4fce5dc8-8114-4ab2-a94b-b4536c27f43b
        type: string
      id_document_url:
        example: www.id-doc-url.com/id.pdf
        type: string
      id_expiry:
        example: "2040-12-31"
        type: string
      id_number:
        example: GHA-123-456-7890
        type: string
      id_type:
        example: GHANACARD
        type: string
      latitude:
        example: 5.6037
        type: number
      logo_url:
        example: www.logo-url.com/logo.png
        type: string
      longitude:
        example: -0.187
        type: number
      name:
        example: Company Name
        type: string
      region:
        example: Greater Accra
        type: string
      registration_number:
        description: company specific fields
        example: GR1234567890
        type: string
      rejected_because:
        example: No reason
        type: string
      rejected_by:
        $ref: '#/definitions/transformations.OutputAdmin'
      rejected_by_id:
        example: R234110
        type: string
      status:
        example: ClientApplication.Status.Approved
        type: string
      sub_type:
        example: ESTATE MANAGER
        type: string
      support_email:
        example: support@email.com
        type: string
      support_phone:
        example: +233 (0)12 345 6789
        type: string
      type:
        example: COMPANY
        type: string
      updated_at:
        example: "2023-01-01T00:00:00Z"
        type: string
      website_url:
        example: www.company-url.com
        type: string
    type: object
  transformations.OutputClientUser:
    properties:
      client:
        $ref: '#/definitions/transformations.OutputClient'
      client_id:
        example: e4ad26d4-d7e9-4599-a246-5e88abba6083
        type: string
      created_at:
        example: "2023-01-01T00:00:00Z"
        type: string
      created_by_id:
        example: 0205126b-9bbb-4a98-960a-e87d8f095335
        type: string
      email:
        example: client-user@example.com
        type: string
      id:
        example: 4fce5dc8-8114-4ab2-a94b-b4536c27f43b
        type: string
      name:
        example: Client User Name
        type: string
      phone_number:
        example: "+233281234569"
        type: string
      role:
        example: STAFF
        type: string
      status:
        example: ClientUser.Status.Active
        type: string
      status_updated_by:
        $ref: '#/definitions/transformations.OutputClientUser'
      status_updated_by_id:
        example: 0205126b-9bbb-4a98-960a-e87d8f095335
        type: string
      updated_at:
        example: "2023-01-01T00:00:00Z"
        type: string
    type: object
...
  transformations.OutputProperty:
    properties:
      address:
        example: 123 Main St
        type: string
      city:
        example: Accra
        type: string
      client:
        $ref: '#/definitions/transformations.OutputClient'
      client_id:
        example: b50874ee-1a70-436e-ba24-572078895982
        type: string
      country:
        example: Ghana
        type: string
      created_at:
        example: "2023-01-01T00:00:00Z"
        format: date-time
        type: string
      created_by:
        $ref: '#/definitions/transformations.OutputClientUser'
      created_by_id:
        example: 1e81fea0-5e8b-4535-b449-1a2133e94a7a
        type: string
      description:
        example: Very elegant place
        type: string
      gps_address:
        example: GH-1234-5678
        type: string
      id:
        example: 4fce5dc8-8114-4ab2-a94b-b4536c27f43b
        format: uuid
        type: string
      images:
        example:
        - http://www.images/hih.jpg
        items:
          type: string
        type: array
      latitude:
        example: 5.6037
        type: number
      longitude:
        example: -0.187
        type: number
      name:
        example: My Property
        type: string
      region:
        example: Greater Accra
        type: string
      slug:
        example: my-property-abcde1876drkjy
        type: string
      status:
        example: Property.Status.Active
        type: string
      tags:
        example:
        - apartment
        - downtown
        items:
          type: string
        type: array
      type:
        example: SINGLE
        type: string
      updated_at:
        example: "2023-01-01T00:00:00Z"
        format: date-time
        type: string
    type: object
  transformations.OutputPropertyBlock:
    properties:
      created_at:
        example: "2023-01-01T00:00:00Z"
        format: date-time
        type: string
      description:
        example: Very elegant place
        type: string
      floors_count:
        example: 3
        type: integer
      id:
        example: 4fce5dc8-8114-4ab2-a94b-b4536c27f43b
        format: uuid
        type: string
      images:
        example:
        - https://example.com/image1.jpg
        - https://example.com/image2.jpg
        items:
          type: string
        type: array
      name:
        example: Block A
        type: string
      property:
        $ref: '#/definitions/transformations.OutputProperty'
      property_id:
        example: 4fce5dc8-8114-4ab2-a94b-b4536c27f43b
        format: uuid
        type: string
      status:
        example: PropertyBlock.Status.Active
        type: string
      units_count:
        example: 3
        type: integer
      updated_at:
        example: "2023-01-01T00:00:00Z"
        format: date-time
        type: string
    type: object
...
info:
  contact:
    email: domeybenjamin1@gmail.com
    name: Domey Benjamin
    url: https://www.linkedin.com/in/benjamin-armah-kesey-domey/
  description: This is the RentLoop API documentation.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://rentloop.com/terms
  title: RentLoop API
  version: "1.0"
paths:
...
  /api/v1/admin/properties/{property_id}/blocks:
    get:
      consumes:
      - application/json
      description: List property blocks (Admin)
      parameters:
      - description: Property ID
        in: path
        name: property_id
        required: true
        type: string
      - in: query
        name: end_date
        type: string
      - collectionFormat: multi
        example:
        - a8098c1a-f86e-11da-bd1a-00112444be1e
        in: query
        items:
          type: string
        name: ids
        type: array
      - enum:
        - asc
        - desc
        in: query
        name: order
        type: string
      - in: query
        name: order_by
        type: string
      - in: query
        minimum: 0
        name: page
        type: integer
      - in: query
        minimum: 0
        name: page_size
        type: integer
      - collectionFormat: csv
        in: query
        items:
          type: string
        name: populate
        type: array
      - in: query
        name: query
        type: string
      - collectionFormat: csv
        in: query
        items:
          type: string
        minItems: 1
        name: search_fields
        type: array
      - in: query
        name: start_date
        type: string
      - enum:
        - PropertyBlock.Status.Active
        - PropertyBlock.Status.Inactive
        - PropertyBlock.Status.Maintenance
        example: PropertyBlock.Status.Active
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                properties:
                  meta:
                    $ref: '#/definitions/lib.HTTPReturnPaginatedMetaResponse'
                  rows:
                    items:
                      $ref: '#/definitions/transformations.OutputPropertyBlock'
                    type: array
                type: object
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/lib.HTTPError'
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: List property blocks (Admin)
      tags:
      - PropertyBlocks
    post:
      consumes:
      - application/json
      description: Create a new property block (Admin)
      parameters:
      - description: Property ID
        in: path
        name: property_id
        required: true
        type: string
      - description: Property block details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.CreatePropertyBlockRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Property block created successfully
          schema:
            properties:
              data:
                $ref: '#/definitions/transformations.OutputPropertyBlock'
            type: object
        "400":
          description: Error occurred when creating a property block
          schema:
            $ref: '#/definitions/lib.HTTPError'
        "401":
          description: Invalid or absent authentication token
          schema:
            type: string
        "403":
          description: Forbidden access
          schema:
            $ref: '#/definitions/lib.HTTPError'
        "422":
          description: Invalid request body
          schema:
            $ref: '#/definitions/lib.HTTPError'
        "500":
          description: An unexpected error occurred
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Create a new property block (Admin)
      tags:
      - PropertyBlocks
  /api/v1/admin/properties/{property_id}/blocks/{block_id}:
    delete:
      consumes:
      - application/json
      description: Delete a property block (Admin)
      parameters:
      - description: Property ID
        in: path
        name: property_id
        required: true
        type: string
      - description: Property block ID
        in: path
        name: block_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: Property block deleted successfully
        "400":
          description: Error occurred when deleting a property block
          schema:
            $ref: '#/definitions/lib.HTTPError'
        "401":
          description: Invalid or absent authentication token
          schema:
            type: string
        "403":
          description: Forbidden access
          schema:
            $ref: '#/definitions/lib.HTTPError'
        "404":
          description: Property block not found
          schema:
            $ref: '#/definitions/lib.HTTPError'
        "500":
          description: An unexpected error occurred
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete a property block (Admin)
      tags:
      - PropertyBlocks
    get:
      consumes:
      - application/json
      description: Get property block by ID (Admin)
      parameters:
      - description: Property ID
        in: path
        name: property_id
        required: true
        type: string
      - description: Property block ID
        in: path
        name: block_id
        required: true
        type: string
      - collectionFormat: csv
        in: query
        items:
          type: string
        name: populate
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: Property block retrieved successfully
          schema:
            properties:
              data:
                $ref: '#/definitions/transformations.OutputPropertyBlock'
            type: object
        "400":
          description: Error occurred when fetching a property block
          schema:
            $ref: '#/definitions/lib.HTTPError'
        "401":
          description: Invalid or absent authentication token
          schema:
            type: string
        "404":
          description: Property block not found
          schema:
            $ref: '#/definitions/lib.HTTPError'
        "500":
          description: An unexpected error occured
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get property block by ID (Admin)
      tags:
      - PropertyBlocks
    patch:
      consumes:
      - application/json
      description: Update a property block (Admin)
      parameters:
      - description: Property ID
        in: path
        name: property_id
        required: true
        type: string
      - description: Property block ID
        in: path
        name: block_id
        required: true
        type: string
      - description: Property block details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdatePropertyBlockRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Property block updated successfully
          schema:
            properties:
              data:
                $ref: '#/definitions/transformations.OutputPropertyBlock'
            type: object
        "400":
          description: Error occurred when updating a property block
          schema:
            $ref: '#/definitions/lib.HTTPError'
        "401":
          description: Invalid or absent authentication token
          schema:
            type: string
        "403":
          description: Forbidden access
          schema:
            $ref: '#/definitions/lib.HTTPError'
        "404":
          description: Property block not found
          schema:
            $ref: '#/definitions/lib.HTTPError'
        "422":
          description: Invalid request body
          schema:
            $ref: '#/definitions/lib.HTTPError'
        "500":
          description: An unexpected error occurred
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Update a property block (Admin)
      tags:
      - PropertyBlocks
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"